transaction updated

+++ credit-card-system/lib/src/main/java/com/cg/creditcardsystem/service/TransactionService.java
@@ -4,6 +4,6 @@ import com.cg.creditcardsystem.dto.TransactionDto;
 import com.cg.creditcardsystem.entities.Transaction;
 
 public interface TransactionService {
-	int addTransaction(TransactionDto transdto);
+	long addTransaction(TransactionDto transdto);
 
 }
